unified jump for playable creatures
